[APP-1003] Moved base app object model classes from search response to aptoide-network module
