+++ aptoide-network/src/main/java/cm/aptoide/pt/aptoide_network/data/network/AppCoins.kt
@@ -1,0 +1,3 @@
+package cm.aptoide.pt.aptoide_network.data.network
+
+data class AppCoins(val advertising: Boolean, val billing: Boolean)
