+++ aptoide-network/src/main/java/cm/aptoide/pt/aptoide_network/data/network/File.kt
@@ -1,0 +1,3 @@
+package cm.aptoide.pt.aptoide_network.data.network
+
+data class File(val malware: Malware)
