+++ aptoide-network/src/main/java/cm/aptoide/pt/aptoide_network/data/network/Malware.kt
@@ -1,0 +1,5 @@
+package cm.aptoide.pt.aptoide_network.data.network
+
+data class Malware(
+  val rank: String
+)
