+++ aptoide-network/src/main/java/cm/aptoide/pt/aptoide_network/data/network/Rating.kt
@@ -1,0 +1,3 @@
+package cm.aptoide.pt.aptoide_network.data.network
+
+data class Rating(val avg: Double, val total: Double)
