+++ aptoide-network/src/main/java/cm/aptoide/pt/aptoide_network/data/network/Stats.kt
@@ -1,0 +1,3 @@
+package cm.aptoide.pt.aptoide_network.data.network
+
+data class Stats(val downloads: Int, val pdownloads: Int, val rating: Rating, val prating: Rating)
