+++ feature_search/src/main/java/cm/aptoide/pt/feature_search/data/fake/FakeRemoteSearchRepository.kt
@@ -1,9 +1,9 @@
 package cm.aptoide.pt.feature_search.data.fake
 
-import cm.aptoide.pt.aptoide_network.data.network.BaseV7DataListResponse
-import cm.aptoide.pt.aptoide_network.data.network.DataList
+import cm.aptoide.pt.aptoide_network.data.network.*
 import cm.aptoide.pt.feature_search.data.network.RemoteSearchRepository
-import cm.aptoide.pt.feature_search.data.network.model.*
+import cm.aptoide.pt.feature_search.data.network.model.SearchAppJsonList
+import cm.aptoide.pt.feature_search.data.network.model.TopSearchAppJsonList
 import cm.aptoide.pt.feature_search.data.network.response.SearchAutoCompleteSuggestionsResponse
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
